move price.ts back into utils
