+++ src/custom/hooks/useUSDCPrice/index.ts
@@ -3,7 +3,7 @@ import { useEffect, useMemo, useState } from 'react'
 import { unstable_batchedUpdates as batchedUpdate } from 'react-dom'
 import { supportedChainId } from 'utils/supportedChainId'
 import { useActiveWeb3React } from 'hooks/web3'
-import { getBestPrice } from 'priceApi/price'
+import { getBestPrice } from '@src/custom/utils/price'
 import { STABLECOIN_AMOUNT_OUT as STABLECOIN_AMOUNT_OUT_UNI } from 'hooks/useUSDCPrice'
 import { stringToCurrency } from 'state/swap/extension'
 import { SupportedChainId } from 'constants/chains'
